requirements.txt
django==5.2.2
psycopg2-binary==2.9.9
django-ninja==1.4.3
django-environ==0.12.0
ipython==9.3.0
akshare==1.16.98